fix: 修复qemu 门禁环境无法退出
Close #I4EAP4

+++ common/qemu_mini_test_monitor.sh
@@ -23,9 +23,21 @@ qemu_count=0
 
 rm -f $test_file
 
+function kill_specified_process(){
+    qemu_name=$1
+    while true
+    do
+        pid=`ps -ef | grep $qemu_name | grep -v grep | awk '{print $2}' | head -n 1`
+        if [ "$pid" == "" ]; then
+            break
+        fi
+        echo "kill pid $pid"
+        kill -15 $pid
+    done
+}
+
 function kill_qemu() {
-    pid=`ps -a | grep qemu-system- | awk '{print $1}'`
-    kill -15 $pid
+    kill_specified_process qemu-system-
     exit 0
 }
